LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/back-forward-cache/eligibility/shared-worker-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Invalid URL

Harness Error (TIMEOUT), message = null

TIMEOUT Eligibility: shared workers Test timed out
